+++ src/main/java/chess/model/piece/Bishop.java
@@ -1,7 +1,7 @@
-package chess.piece;
+package chess.model.piece;
 
 import chess.Direction;
-import chess.square.Square;
+import chess.model.square.Square;
 import java.util.List;
 import java.util.stream.Collectors;
