+++ src/main/java/chess/model/piece/Empty.java
@@ -1,6 +1,6 @@
-package chess.piece;
+package chess.model.piece;
 
-import chess.square.Square;
+import chess.model.square.Square;
 
 public class Empty extends Piece {
