+++ src/main/java/chess/model/piece/King.java
@@ -1,6 +1,6 @@
-package chess.piece;
+package chess.model.piece;
 
-import chess.square.Square;
+import chess.model.square.Square;
 
 public class King extends Piece {
