+++ src/test/java/chess/model/square/FileTest.java
@@ -1,8 +1,9 @@
-package chess.square;
+package chess.model.square;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.*;
 
+import chess.model.square.File;
 import org.junit.jupiter.api.Test;
 
 class FileTest {
