+++ src/test/java/chess/model/square/SquareTest.java
@@ -1,4 +1,4 @@
-package chess.square;
+package chess.model.square;
 
 import static chess.model.piece.Fixtures.A7;
 import static chess.model.piece.Fixtures.B6;
@@ -7,6 +7,9 @@ import static chess.model.piece.Fixtures.D4;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import chess.Direction;
+import chess.model.square.File;
+import chess.model.square.Rank;
+import chess.model.square.Square;
 import java.util.List;
 import org.junit.jupiter.api.Test;
